library/requirements.txt
django>=4.2
djangorestframework>=3.14
django-filter>=23.2
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21